+++ test/bulkhead_lm_test_connectors_signed_interactive.ml
@@ -40,7 +40,7 @@ let viber_connector_handles_text_webhook_test _switch () =
       ~provider_factory:(fun _ -> provider)
       (Bulkhead_lm.Config_test_support.sample_config
          ~user_connectors:
-           { Bulkhead_lm.Config.telegram = None
+           { Bulkhead_lm.Config.telegram = []
            ; whatsapp = None
            ; messenger = None
            ; instagram = None
@@ -181,7 +181,7 @@ let wechat_connector_handles_encrypted_verification_test _switch () =
     Bulkhead_lm.Runtime_state.create
       (Bulkhead_lm.Config_test_support.sample_config
          ~user_connectors:
-           { Bulkhead_lm.Config.telegram = None
+           { Bulkhead_lm.Config.telegram = []
            ; whatsapp = None
            ; messenger = None
            ; instagram = None
@@ -296,7 +296,7 @@ let discord_connector_handles_command_webhook_test _switch () =
       ~provider_factory:(fun _ -> provider)
       (Bulkhead_lm.Config_test_support.sample_config
          ~user_connectors:
-           { Bulkhead_lm.Config.telegram = None
+           { Bulkhead_lm.Config.telegram = []
            ; whatsapp = None
            ; messenger = None
            ; instagram = None