+++ test/bulkhead_lm_test_connectors_social_text.ml
@@ -43,7 +43,7 @@ let instagram_connector_handles_text_webhook_test _switch () =
       ~provider_factory:(fun _ -> provider)
       (Bulkhead_lm.Config_test_support.sample_config
          ~user_connectors:
-           { Bulkhead_lm.Config.telegram = None
+           { Bulkhead_lm.Config.telegram = []
            ; whatsapp = None
            ; messenger = None
            ; instagram = Some connector
@@ -221,7 +221,7 @@ let line_connector_handles_text_webhook_test _switch () =
       ~provider_factory:(fun _ -> provider)
       (Bulkhead_lm.Config_test_support.sample_config
          ~user_connectors:
-           { Bulkhead_lm.Config.telegram = None
+           { Bulkhead_lm.Config.telegram = []
            ; whatsapp = None
            ; messenger = None
            ; instagram = None
@@ -363,7 +363,7 @@ let wechat_connector_handles_verification_test _switch () =
     Bulkhead_lm.Runtime_state.create
       (Bulkhead_lm.Config_test_support.sample_config
          ~user_connectors:
-           { Bulkhead_lm.Config.telegram = None
+           { Bulkhead_lm.Config.telegram = []
            ; whatsapp = None
            ; messenger = None
            ; instagram = None